fix compatibility issue on CentOS

+++ install.sh
@@ -5,7 +5,7 @@ set -e
 SCRIPT_DIR="$(cd $(dirname $0) && pwd -L)"
 BACKUP_DIR="$HOME/.myrc/backup/$(date +%Y-%m-%d)"
 HOME_DIR="$HOME"
-RCFILES="$(find "$SCRIPT_DIR" -name ".*" -type f -depth 1 | xargs basename)"
+RCFILES="$(find "$SCRIPT_DIR" -maxdepth 1 -name ".*" -type f | xargs -I '{}' basename {})"
 
 for ((i=0; ; i++)); do
     if [ $i = "0" ] && [ ! -e "$BACKUP_DIR" ]; then
@@ -21,7 +21,9 @@ mkdir -p "$DEST_DIR"
 
 # Back up old files.
 for f in $RCFILES; do
-    cp -pLR "$HOME_DIR/$f" "$DEST_DIR/"
+    if [ -e "$HOME_DIR/$f" ]; then
+        cp -pLR "$HOME_DIR/$f" "$DEST_DIR/"
+    fi
 done
 
 # Link all rc files.
